Change variable type in BookDto from double to float

+++ src/main/java/com/makowski/bookshelves/dto/BookDto.java
@@ -11,6 +11,7 @@ public class BookDto {
     private String title;
     private String author;
     private int year;
-    private double rating;
+    private float rating;
     private int scoresNumber;
+
 }
